requirements.txt
gpiozero~=1.6.2
omxplayer-wrapper~=0.3.3
pygame~=2.1.0